server/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
